isky-audit-mservice/requirements.txt
absl-py==1.0.0
amqp==5.1.0
appwrite==1.1.0
astunparse==1.6.3
attrs==22.1.0
bcrypt==4.0.1
beautifulsoup4==4.10.0
billiard==3.6.4.0
cachetools==5.0.0
celery==5.2.3
certifi==2021.10.8
cffi==1.15.1
charset-normalizer==2.0.11
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.4
cryptography==38.0.2
cytoolz==0.11.0
dask==2022.2.1
dependency-injector==4.40.0
fasteners==0.17.3
filterpy==1.4.5
Flask==2.2.2
Flask-Bcrypt==1.0.1
Flask-Cors==3.0.10
Flask-JWT-Extended==4.4.4
flask-openapi3==2.1.0
flatbuffers==2.0
fonttools==4.25.0
fsspec==2022.2.0
gast==0.5.3
gunicorn==20.1.0
humanize==4.0.0
idna==3.3
importlib-metadata==4.11.1
iniconfig==1.1.1
itsdangerous==2.1.2
kombu==5.2.4
libclang==13.0.0
locket==0.2.1
Markdown==3.3.6
MarkupSafe==2.1.1
munkres==1.1.4
nose==1.3.7
oauthlib==3.2.0
opt-einsum==3.3.0
pluggy==1.0.0
prometheus-client==0.13.1
prompt-toolkit==3.0.28
protobuf==3.19.4
py==1.11.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydantic==1.9.1
PyJWT==2.5.0
pyOpenSSL==22.1.0
pyoselm==1.0.1
pytest==7.1.3
python-dotenv==0.21.0
pytz==2021.3
PyYAML==6.0
requests==2.27.1
requests-oauthlib==1.3.1
rethinkdb==2.4.9
rsa==4.8
soupsieve==2.3.1
tables==3.7.0
tabulate==0.8.10
termcolor==1.1.0
tomli==2.0.1
tqdm==4.62.3
typer==0.4.1
typing_extensions==4.1.1
urllib3==1.26.8
vine==5.0.0
wcwidth==0.2.5
Werkzeug==2.2.2
wrapt==1.13.3
zict==2.0.0
zipp==3.7.0
